#87 continue fix tasklist follow fontsize
